tests/06-inherit-layers-from-attribute.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Replace Layer Statement Rule</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>
    <style>
        @layer BETTER-BUTTONS {
            button {
                padding: 1em;
                border: thick dashed blue;
            }
        }
    </style>
    <custom-element>
        <template shadowrootmode="open">
            <style>
                @layer buttons;

                @layer buttons {
                    button {
                        padding: 1em;
                        border: 1px solid #000;
                    }
                }
            </style>
            <button>custom element with " buttons " layer</button>
        </template>
    </custom-element>

    <custom-element-with-attribute shadowlayers="buttons, inherit.BETTER-BUTTONS">
        <template shadowrootmode="open">
            <style>
                @layer buttons;

                @layer buttons {
                    button {
                        padding: 1em;
                        border: 1px solid #000;
                    }
                }
            </style>
            <button>custom element with "buttons inherit.BETTER-BUTTONS" layers
            </button>
        </template>
    </custom-element-with-attribute>


    <script type="module">
        import ShadowLayers from '../ShadowLayers.js';
        import { adoptStylesFromAttribute } from "../AdoptStyles.js";

        window.onload = function () {

            const elName = "custom-element-with-attribute"
            const newLayers = document.querySelector(elName).getAttribute("shadowlayers").split(", ")

            test(function () {

                const inheritFromAttributeResult = adoptStylesFromAttribute(document.querySelector(elName))

                assert_true(inheritFromAttributeResult, "inheritFromAttribute() successfully returns");

                const layerStatementResult = ShadowLayers.findLayerStatementRule(document.querySelector(elName))
                assert_equals(typeof layerStatementResult, "object", "found, so returns object");

                const layerStatementRule = layerStatementResult.styleSheet.cssRules.item(layerStatementResult.ruleIndex)
                assert_array_equals(newLayers,
                    layerStatementRule.nameList,
                    "new @layer = buttons inherit.BETTER-BUTTONS");

            }, "Inherit inherit.BETTER-BUTTONS layer from page styles");

        }
    </script>

    <div id="log"></div>

</body>

</html>
